Pull validations-related methods out into a separate module
